feat(venn): SJIP-1254 extract venn chart
